film-community/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
neo4j>=5.14.0
python-dotenv>=1.0.0

